Updates RacingGameProject Struture
