+++ src/main/java/model/RacingGame.java
@@ -1,0 +1,17 @@
+package model;
+
+import controller.CarController;
+import view.InputView;
+
+import java.util.*;
+
+public class RacingGame {
+
+    private int numberOfIterations;
+    private List<CarController> carControllerList;
+
+    public RacingGame(int numberOfIterations, List<CarController> carControllerList) {
+        this.numberOfIterations = numberOfIterations;
+        this.carControllerList = carControllerList;
+    }
+}
